1.10 JEI improvements
 - Bump JEI version
 - Add icons for upgrade handlers
 - Add custom subtype comparisons for turtles & pocket computers

+++ src/main/java/org/squiddev/cctweaks/integration/ForgeIntegration.java
@@ -39,7 +39,7 @@ public class ForgeIntegration extends Module {
 					change = producer.extractEnergy(fluxLimit - progress, false);
 					progress += change;
 				}
-				return (int) (progress / fluxAmount);
+				return progress / fluxAmount;
 			}
 		});
 	}
